+++ packages/pancake/CHANGELOG.md
@@ -6,6 +6,7 @@ Pancake
 
 ## Versions
 
+* [v1.1.1  - Optimized pancake for yarn monorepos](v111)
 * [v1.1.0  - Optimized pancake for windows](v110)
 * [v1.0.15 - Fixed mixed js and Sass module bug](v1015)
 * [v1.0.14 - Tweaked to facilitate syrup](v1013)
@@ -22,6 +23,10 @@ Pancake
 
 ----------------------------------------------------------------------------------------------------------------------------------------------------------------
 
+## v1.1.1
+
+- Pancake now searches the folder above the current/package.json folder to better support monorepos via yarn
+	([#15](https://github.com/govau/pancake/issues/15))
 
 ## v1.1.0